Refactor Rect interface moved to util
Add calculateBoundingRect function
